Address review feedback on accelerated phases
- Add accelerated_phase_window_seconds to GameListSerializer (was missing)
- Hoist inline imports in phase/tests.py to top of file per CLAUDE.md
- Extract FREQUENCY_SECONDS and formatDurationSeconds to shared
  src/utils/deadline.ts; import from there in DeadlineSummary and CreateGame
- Delete one-off screenshot-interact.mjs helper script
- Add missing <FormMessage /> to acceleratedPhaseEnabled checkbox FormField
- Revert activeGameRetreat fixture to its original state (no fixed_time
  or acceleration data); window-closed scenario covered by the new
  activeGameRetreatWindowClosed fixture
