+++ packages/web/src/components/DeadlineSummary.tsx
@@ -1,26 +1,5 @@
 import React from "react";
-
-const FREQUENCY_SECONDS: Record<string, number> = {
-  hourly: 3600,
-  daily: 86400,
-  every_2_days: 172800,
-  weekly: 604800,
-};
-
-function formatDurationSeconds(seconds: number): string {
-  if (seconds < 3600) {
-    const mins = Math.round(seconds / 60);
-    return `${mins} ${mins === 1 ? "minute" : "minutes"}`;
-  }
-  if (seconds % 86400 === 0) {
-    const days = seconds / 86400;
-    return `${days} ${days === 1 ? "day" : "days"}`;
-  }
-  const hrs = Math.floor(seconds / 3600);
-  const mins = Math.round((seconds % 3600) / 60);
-  if (mins === 0) return `${hrs} ${hrs === 1 ? "hour" : "hours"}`;
-  return `${hrs}h ${mins}m`;
-}
+import { FREQUENCY_SECONDS, formatDurationSeconds } from "@/utils/deadline";
 
 interface DeadlineSummaryGame {
   movementPhaseDuration: string | null;
